Changed cart items to be stored in session instead of the database.

+++ Shop.Application/Cart/CartDto.cs
@@ -11,7 +11,7 @@ namespace Shop.Application.Cart
     public class CartDto
     {
         public string Id { get; set; }
-        public float CartTotal { get; set; } // suma zamówienia
+        public decimal CartTotal { get; set; } // suma zamówienia
         public List<CartItem> CartItems { get; set; } 
     }
 }
